Cleanup ClusterHttpClientUtil
* Rename to TopologyChangeListenerUtil
* Remove legacy unused tryGet and tryGetAndConsume methods
* Use try-with-resources pattern for CloseableHttpClient and  CloseableHttpResponse
* Fix class javadoc
