+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/web/AbstractWebFailoverTestCase.java
@@ -27,7 +27,7 @@ import org.infinispan.transaction.TransactionMode;
 import org.jboss.arquillian.container.test.api.OperateOnDeployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
-import org.jboss.as.test.clustering.ClusterHttpClientUtil;
+import org.jboss.as.test.clustering.TopologyChangeListenerUtil;
 import org.jboss.as.test.clustering.cluster.AbstractClusteringTestCase;
 import org.jboss.as.test.clustering.single.web.SimpleServlet;
 import org.jboss.as.test.http.util.TestHttpClientUtils;
@@ -374,7 +374,7 @@ public abstract class AbstractWebFailoverTestCase extends AbstractClusteringTest
 
     private void establishTopology(URL baseURL, Set<String> topology) throws URISyntaxException, IOException, InterruptedException {
         if (this.cacheMode.isClustered()) {
-            ClusterHttpClientUtil.establishTopology(baseURL, "web", this.deploymentName, topology);
+            TopologyChangeListenerUtil.establishTopology(baseURL, "web", this.deploymentName, topology);
 
             // TODO we should be able to speed this up by observing changes in the routing registry
             // prevents failing assertions when topology information is expected, e.g.:
